requirements.txt
Django==2.1.4
django-cors-headers==2.4.0
django-rest-auth==0.9.3
djangorestframework==3.9.0
docker==3.7.0
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
psycopg2==2.7.4
python-dateutil==2.6.0
urllib3==1.24.1
websocket==0.2.1
websocket-client==0.54.0
